splited files into orginised folders
